login

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
@@ -7,7 +9,10 @@ from .models import Room, Topic
 from .forms import RoomForm
 
 def loginPage(request):
-    
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -25,13 +30,18 @@ def loginPage(request):
         else:
             messages.error(request, 'Nome de usuário ou senha incorreto')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+
+def registerPage(request):
+    page = 'register'
+    return render(request, 'base/login_register.html')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -48,11 +58,14 @@ def home(request):
     context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, "base/home.html", context)
 
+
 def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {'room':room}
     return render(request, "base/room.html", context)
 
+
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -64,9 +77,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Você não tem acesso a esta sala')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -77,8 +95,14 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room =Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Você não tem acesso a esta sala')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
